removed all methods from xmtp context
