+++ contexts/xmtp.ts
@@ -1,12 +1,5 @@
 import { createContext } from 'react'
-import { Signer } from 'ethers'
 
-export type XmtpContextType = {
-  initClient: (wallet: Signer) => void
-}
-
-export const XmtpContext = createContext<XmtpContextType>({
-  initClient: () => undefined,
-})
+export const XmtpContext = createContext({})
 
 export default XmtpContext
